Add .workrc and improve install script

+++ install.sh
@@ -2,8 +2,13 @@
 for file in dotfiles/.*[a-z]; do
   file_name=$(basename $file)
 
-  rm ~/$file_name
-  
+
+  if [ -e ~/$home_dotfile ] ||\
+       [ -f ~/$home_dotfile ] ||\
+       [ -L ~/$home_dotfile ];then
+    rm ~/$file_name
+  fi
+ 
   ln -s "$(pwd)/$file" ~/$file_name
 done
